added telemetry to namespace
