+++ src/main/java/com/rackspace/salus/telemetry/presence_monitor/PresenceMonitorApplication.java
@@ -1,4 +1,4 @@
-package com.rackspace.salus.presence_monitor;
+package com.rackspace.salus.telemetry.presence_monitor;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
