+++ src/main/java/com/rackspace/salus/telemetry/presence_monitor/services/KafkaEgress.java
@@ -16,10 +16,10 @@
  *
  */
 
-package com.rackspace.salus.presence_monitor.services;
+package com.rackspace.salus.telemetry.presence_monitor.services;
 
-import com.rackspace.salus.presence_monitor.config.PresenceMonitorProperties;
-import com.rackspace.salus.presence_monitor.types.KafkaMessageType;
+import com.rackspace.salus.telemetry.presence_monitor.config.PresenceMonitorProperties;
+import com.rackspace.salus.telemetry.presence_monitor.types.KafkaMessageType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
