+++ src/main/java/com/rackspace/salus/telemetry/presence_monitor/services/MetricExporter.java
@@ -1,7 +1,7 @@
-package com.rackspace.salus.presence_monitor.services;
+package com.rackspace.salus.telemetry.presence_monitor.services;
 
-import com.rackspace.salus.presence_monitor.types.PartitionEntry;
-import com.rackspace.salus.presence_monitor.config.PresenceMonitorProperties;
+import com.rackspace.salus.telemetry.presence_monitor.types.PartitionEntry;
+import com.rackspace.salus.telemetry.presence_monitor.config.PresenceMonitorProperties;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
