+++ src/main/java/com/rackspace/salus/telemetry/presence_monitor/types/KafkaMessageType.java
@@ -16,19 +16,10 @@
  *
  */
 
-package com.rackspace.salus.presence_monitor.config;
+package com.rackspace.salus.telemetry.presence_monitor.types;
 
-import com.rackspace.salus.presence_monitor.types.KafkaMessageType;
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-import java.util.Map;
-
-@ConfigurationProperties("presence-monitor")
-@Component
-@Data
-public class    PresenceMonitorProperties {
-    Map<KafkaMessageType, String> kafkaTopics;
-    int exportPeriodInSeconds;
+public enum KafkaMessageType {
+    LOG,
+    METRIC,
+    EVENT;
 }
